Fix the test
