+++ frontend/src/metabase-types/api/settings.ts
@@ -1,6 +1,11 @@
+export interface Version {
+  tag: string;
+}
+
 export interface Settings {
   "enable-public-sharing"?: boolean;
   "enable-xrays"?: boolean;
+  "engine-deprecation-notice-version"?: string;
   "show-database-syncing-modal"?: boolean;
   "show-homepage-data"?: boolean;
   "show-homepage-xrays"?: boolean;
@@ -9,4 +14,5 @@ export interface Settings {
   "slack-token-valid?"?: boolean;
   "slack-app-token"?: string;
   "slack-files-channel"?: string;
+  version?: Version;
 }
