+++ frontend/src/metabase/admin/app/components/AdminApp/index.ts
