+++ frontend/src/metabase/admin/app/containers/DeprecationBanner/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DeprecationBanner";
