--- frontend/src/metabase/admin/app/containers/DriverDeprecationBanner/index.js
@@ -1,1 +1,0 @@
-export { default } from "./DriverDeprecationBanner";
